+++ includes/SpectroCoinBlocksIntegration.php
@@ -1,28 +1,29 @@
 <?php
 
+declare(strict_types=1);
+
 namespace SpectroCoin\Includes;
 
 use SpectroCoin\Includes\SpectroCoinGateway;
-
 use Automattic\WooCommerce\Blocks\Payments\Integrations\AbstractPaymentMethodType;
 
 if (!defined('ABSPATH')) {
-	die('Access denied.');
+    die('Access denied.');
 }
 
 final class SpectroCoinBlocksIntegration extends AbstractPaymentMethodType {
-    private $gateway;
+    private SpectroCoinGateway $gateway;
     protected $name = 'spectrocoin';
 
-    public function initialize() {
+    public function initialize(): void {
         $this->gateway = new SpectroCoinGateway();
     }
 
-    public function is_active() {
+    public function is_active(): bool {
         return $this->gateway->is_available();
     }
 
-    public function get_payment_method_script_handles() {
+    public function get_payment_method_script_handles(): array {
         wp_register_script(
             'spectrocoin-blocks-integration',
             plugin_dir_url(__FILE__) . '../assets/js/block-checkout.js',
@@ -38,7 +39,7 @@ final class SpectroCoinBlocksIntegration extends AbstractPaymentMethodType {
         return ['spectrocoin-blocks-integration'];
     }
 
-    public function get_payment_method_data() {
+    public function get_payment_method_data(): array {
         $checkout_icon_url = plugins_url('/assets/images/spectrocoin-logo.svg', __DIR__);
         return [
             'title' => $this->gateway->title,