+++ changelog.md
@@ -36,3 +36,5 @@ _For Developers_: Added documentation with parameters and return variables befor
 _Fixed_: Replaced hardcoded order statuses in plugin settings.
 
 _Added_: Custom order statuses created manually or using plugins will appear in SpectroCoin settings menu.
+
+_Added_: During checkout, if error is occured, now client will see the error code and message instead of generic error message.