+++ changelog.md
@@ -40,3 +40,5 @@ _Added_: Custom order statuses created manually or using plugins will appear in 
 _Added_: During checkout, if error is occured, now client will see the error code and message instead of generic error message.
 
 _Added:_ Now plugin checks the FIAT currency, if it is not supported by SpectroCoin, payment will not be available.
+
+_Added:_ Added admin notice in admin plugin settings to notify that shop currency is not supported by SpectroCoin.